Create watch command
The watch command uses the watchdog package to compile the paper
whenever a source file is changed.

Extension names to look for can be specified, as can the name of the
command to be run (typically `make`).

The global setting `--master` allows the root tex document to be
specified, and thus let the watch command ignore changes to the compiled
PDF.