Create a configuration system
There is a Configurations class that manages defaults. Configurations
can be set from internal defaults, but overriden by a `preprint.json`
file that can be added to a project's directory. The settings in
Configurations are only used to populate argparser defaults. Each
command uses the result of argparser to get the definitive command
configs.