Create latexdiff command
This command is a pipeline for running latexdiff between the current
manuscript, and a revision of the manuscript in git. Any SHA fragment
can be passed. GitPython is used to read git blobs.

Documents are are included with \input{} commands are recursively
inlined to support latexdiff.

The command then runs latexdiff and latexmk to compile the difference
document.

Updated README for preprint diff