Add inline.py module
The inline function uses regular expression matching \input{*}
expressions, and inlines those text files into the root text. This
inlining is rescursive. We use Python's re.sub, with a function
processing each match.

All files are assumed to be UTF-8 unicode files. File opening is
done with codecs.open. It could be a good idea to give the user a choice
of codec.